requirements.txt
Django==5.1.3
django-ninja==1.3.0
google-generativeai==0.8.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.4.0
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==41.0.7
